Revert "Improve Anime4K quality profile"
This reverts commit 83991e9c6dd8ba8085d9eaa79ad15fcc02e7f1b9.
